Fixed compiler warnings.
git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1705 d9297710-12f9-482b-b11f-fb677dc88bbb

+++ src/io/discretization/Peridigm_AlbanyDiscretization.cpp
@@ -206,7 +206,7 @@ void PeridigmNS::AlbanyDiscretization::createBlockElementLists() {
   // Find the unique block ids on processor
   set<int> uniqueBlockIds;
   for(int i=0 ; i<blockID->MyLength() ; ++i)
-    uniqueBlockIds.insert((*blockID)[i]);
+    uniqueBlockIds.insert( static_cast<int>( (*blockID)[i] ) );
 
   // Broadcast the unique block ids so that all processors are aware of the full block list
   Teuchos::RCP<const Teuchos::Comm<int> > teuchosComm = Teuchos::createMpiComm<int>(Teuchos::opaqueWrapper<MPI_Comm>(MPI_COMM_WORLD));
